Add files via upload
This is a simple TO-DO list i created and here are the features of the simple project, what you can achieve and and the languages i used in developing it.
Features: Ablity to add tasks, edit tasks, delete tasks, set deadlines, categorize tasks, and much more.
Technology behind this project: HTML, CSS and javascript only though with the use of a text editor known as sublime text....That sounds unreal right? But that's the truth.
Trust me it's simple to use and self explanatory.
Always open for any corrections, critize, compliment and much more and if any, please make a pull request.
Good luck to all young developers.

See you all.